Add support to Freeze tenant
