+++ pkg/apis/minio.min.io/v1/zz_generated.deepcopy.go
@@ -199,6 +199,11 @@ func (in *TenantSpec) DeepCopyInto(out *TenantSpec) {
 		*out = new(corev1.LocalObjectReference)
 		**out = **in
 	}
+	if in.Freeze != nil {
+		in, out := &in.Freeze, &out.Freeze
+		*out = new(bool)
+		**out = **in
+	}
 	return
 }
