+++ pkg/controller/cluster/log-search.go
@@ -18,8 +18,14 @@
 package cluster
 
 import (
+	"context"
 	"errors"
 	"fmt"
+	"net/http"
+	"time"
+
+	"github.com/minio/madmin-go"
+	"github.com/minio/operator/pkg/resources/secrets"
 
 	miniov2 "github.com/minio/operator/pkg/apis/minio.min.io/v2"
 	"github.com/minio/operator/pkg/resources/deployments"
@@ -28,6 +34,8 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/klog/v2"
 )
 
@@ -36,6 +44,52 @@ type auditWebhookConfig struct {
 	args   string
 }
 
+func (c *Controller) syncLogSearchState(ctx context.Context, tenant *miniov2.Tenant, adminClnt *madmin.AdminClient) error {
+	if tenant.HasLogEnabled() {
+		_, err := c.checkAndCreateLogSecret(ctx, tenant)
+		if err != nil {
+			return err
+		}
+
+		var logSecret *corev1.Secret
+		logSecret, err = c.checkAndCreateLogSecret(ctx, tenant)
+		if err != nil {
+			return err
+		}
+
+		searchSvc, err := c.checkAndCreateLogHeadless(ctx, tenant)
+		if err != nil {
+			return err
+		}
+
+		err = c.checkAndCreateLogStatefulSet(ctx, tenant, searchSvc.Name)
+		if err != nil {
+			return err
+		}
+
+		err = c.checkAndCreateLogSearchAPIDeployment(ctx, tenant)
+		if err != nil {
+			return err
+		}
+
+		err = c.checkAndCreateLogSearchAPIService(ctx, tenant)
+		if err != nil {
+			return err
+		}
+		// Make sure that MinIO is up and running to enable Log Search.
+		if !tenant.MinIOHealthCheck() {
+			if _, err = c.updateTenantStatus(ctx, tenant, StatusWaitingForReadyState, -1); err != nil {
+				return err
+			}
+			return ErrMinIONotReady
+		}
+		err = c.checkAndConfigureLogSearchAPI(ctx, tenant, logSecret, adminClnt)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
 func newAuditWebhookConfig(tenant *miniov2.Tenant, secret *corev1.Secret) auditWebhookConfig {
 	auditToken := string(secret.Data[miniov2.LogAuditTokenKey])
 	whTarget := fmt.Sprintf("audit_webhook:%s", tenant.LogSearchAPIDeploymentName())
@@ -92,4 +146,159 @@ func logSearchAPIDeploymentMatchesSpec(tenant *miniov2.Tenant, actualDeployment 
 		return false, nil
 	}
 	return true, nil
+}
+
+func (c *Controller) checkAndCreateLogHeadless(ctx context.Context, tenant *miniov2.Tenant) (*corev1.Service, error) {
+	svc, err := c.serviceLister.Services(tenant.Namespace).Get(tenant.LogHLServiceName())
+	if err == nil || !k8serrors.IsNotFound(err) {
+		return svc, err
+	}
+
+	klog.V(2).Infof("Creating a new Log Headless Service for %s", tenant.Namespace)
+	svc = services.NewHeadlessForLog(tenant)
+	_, err = c.kubeClientSet.CoreV1().Services(svc.Namespace).Create(ctx, svc, metav1.CreateOptions{})
+	return svc, err
+}
+
+func (c *Controller) checkAndCreateLogStatefulSet(ctx context.Context, tenant *miniov2.Tenant, svcName string) error {
+	logPgSS, err := c.statefulSetLister.StatefulSets(tenant.Namespace).Get(tenant.LogStatefulsetName())
+	if err != nil {
+		if !k8serrors.IsNotFound(err) {
+			return err
+		}
+		if tenant, err = c.updateTenantStatus(ctx, tenant, StatusProvisioningLogPGStatefulSet, 0); err != nil {
+			return err
+		}
+
+		klog.V(2).Infof("Creating a new Log StatefulSet for %s", tenant.Namespace)
+		searchSS := statefulsets.NewForLogDb(tenant, svcName)
+		_, err = c.kubeClientSet.AppsV1().StatefulSets(tenant.Namespace).Create(ctx, searchSS, metav1.CreateOptions{})
+		return err
+
+	}
+
+	// check if expected and actual values of Log DB spec match
+	dbSpecMatches, err := logDBStatefulsetMatchesSpec(tenant, logPgSS)
+	if err != nil {
+		return err
+	}
+	if !dbSpecMatches {
+		// Note: using current spec replica count works as long as we don't expose replicas via tenant spec.
+		if tenant, err = c.updateTenantStatus(ctx, tenant, StatusUpdatingLogPGStatefulSet, *logPgSS.Spec.Replicas); err != nil {
+			return err
+		}
+		logPgSS = statefulsets.NewForLogDb(tenant, svcName)
+		if _, err = c.kubeClientSet.AppsV1().StatefulSets(tenant.Namespace).Update(ctx, logPgSS, metav1.UpdateOptions{}); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (c *Controller) checkAndCreateLogSearchAPIService(ctx context.Context, tenant *miniov2.Tenant) error {
+	_, err := c.serviceLister.Services(tenant.Namespace).Get(tenant.LogSearchAPIServiceName())
+	if err == nil || !k8serrors.IsNotFound(err) {
+		return err
+	}
+
+	klog.V(2).Infof("Creating a new Log Search API Service for %s", tenant.Namespace)
+	svc := services.NewClusterIPForLogSearchAPI(tenant)
+	_, err = c.kubeClientSet.CoreV1().Services(svc.Namespace).Create(ctx, svc, metav1.CreateOptions{})
+	return err
+}
+
+func (c *Controller) checkAndCreateLogSearchAPIDeployment(ctx context.Context, tenant *miniov2.Tenant) error {
+	logSearchDeployment, err := c.deploymentLister.Deployments(tenant.Namespace).Get(tenant.LogSearchAPIDeploymentName())
+	if err != nil {
+		if !k8serrors.IsNotFound(err) {
+			return err
+		}
+		if tenant, err = c.updateTenantStatus(ctx, tenant, StatusProvisioningLogSearchAPIDeployment, 0); err != nil {
+			return err
+		}
+
+		klog.V(2).Infof("Creating a new Log Search API deployment for %s", tenant.Name)
+		_, err = c.kubeClientSet.AppsV1().Deployments(tenant.Namespace).Create(ctx, deployments.NewForLogSearchAPI(tenant), metav1.CreateOptions{})
+		return err
+	}
+
+	// check if expected and actual values of Log search API deployment match
+	apiDeploymentMatches, err := logSearchAPIDeploymentMatchesSpec(tenant, logSearchDeployment)
+	if err != nil {
+		return err
+	}
+	if !apiDeploymentMatches {
+		// Note: using current spec replica count works as long as we don't expose replicas via tenant spec.
+		if tenant, err = c.updateTenantStatus(ctx, tenant, StatusUpdatingLogSearchAPIServer, 0); err != nil {
+			return err
+		}
+		logSearchDeployment = deployments.NewForLogSearchAPI(tenant)
+		if _, err := c.kubeClientSet.AppsV1().Deployments(tenant.Namespace).Update(ctx, logSearchDeployment, metav1.UpdateOptions{}); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (c *Controller) checkAndCreateLogSecret(ctx context.Context, tenant *miniov2.Tenant) (*corev1.Secret, error) {
+	secret, err := c.kubeClientSet.CoreV1().Secrets(tenant.Namespace).Get(ctx, tenant.LogSecretName(), metav1.GetOptions{})
+	if err == nil || !k8serrors.IsNotFound(err) {
+		return secret, err
+	}
+
+	klog.V(2).Infof("Creating a new Log secret for %s", tenant.Name)
+	secret, err = c.kubeClientSet.CoreV1().Secrets(tenant.Namespace).Create(ctx, secrets.LogSecret(tenant), metav1.CreateOptions{})
+	return secret, err
+}
+
+func (c *Controller) checkAndConfigureLogSearchAPI(ctx context.Context, tenant *miniov2.Tenant, secret *corev1.Secret, adminClnt *madmin.AdminClient) error {
+	// Check if audit webhook is configured for tenant's MinIO
+	auditCfg := newAuditWebhookConfig(tenant, secret)
+	_, err := adminClnt.GetConfigKV(ctx, auditCfg.target)
+	if err != nil {
+		// check if log search is ready
+		if err = c.checkLogSearchAPIReady(tenant); err != nil {
+			klog.V(2).Info(err)
+			if _, err = c.updateTenantStatus(ctx, tenant, StatusWaitingForLogSearchReadyState, 0); err != nil {
+				return err
+			}
+			return ErrLogSearchNotReady
+		}
+		restart, err := adminClnt.SetConfigKV(ctx, auditCfg.args)
+		if err != nil {
+			return err
+		}
+		if restart {
+			// Restart MinIO for config update to take effect
+			if err = adminClnt.ServiceRestart(ctx); err != nil {
+				klog.V(2).Info("error restarting minio")
+				klog.V(2).Info(err)
+			}
+			klog.V(2).Info("done restarting minio")
+		}
+		return nil
+	}
+	return err
+}
+
+func (c *Controller) checkLogSearchAPIReady(tenant *miniov2.Tenant) error {
+	endpoint := fmt.Sprintf("http://%s.%s.svc.%s:8080", tenant.LogSearchAPIServiceName(), tenant.Namespace, miniov2.GetClusterDomain())
+	client := http.Client{Timeout: 100 * time.Millisecond}
+	resp, err := client.Get(endpoint)
+	if err != nil {
+		fmt.Println(err)
+		return err
+	}
+	defer func() {
+		if err := resp.Body.Close(); err != nil {
+			klog.V(2).Info(err)
+		}
+	}()
+
+	if resp.StatusCode == 404 {
+		return nil
+	}
+
+	return errors.New("Log Search API Not Ready")
 }
